requirements.txt
# Deep Learning Frameworks
tensorflow>=2.16.0
keras>=3.0.0

# Segmentation Models
segmentation-models>=1.0.1

# Data Processing
numpy>=1.24.0
pandas>=2.0.0
pillow>=10.0.0
opencv-python>=4.8.0

# Geospatial (optional - может потребовать дополнительной настройки)
# rasterio>=1.3.0
# GDAL может быть сложно установить, закомментировано

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0

# Image Processing
scikit-image>=0.21.0

# Machine Learning Utils
scikit-learn>=1.3.0

# Progress Bars
tqdm>=4.65.0

# Data Augmentation for Segmentation
albumentations>=1.3.0

# Jupyter
jupyter>=1.0.0
ipykernel>=6.25.0
notebook>=7.0.0

# Utilities
h5py>=3.9.0
requests>=2.31.0
